Refactor FeedbackAdapter to generic TreeAdapter
