+++ app/src/main/assets/paperwork.json
@@ -1,1 +1,1 @@
-{"buildTime":"2020-09-02 13:59","gitSha":"9f36d3ea5"}
+{"buildTime":"2020-09-02 17:33","gitSha":"0fd670677"}
