+++ app/src/psi/java/org/dhis2/core/ui/tree/TreeAdapterBinder.kt
@@ -1,0 +1,10 @@
+package org.dhis2.core.ui.tree
+
+import android.view.View
+import androidx.recyclerview.widget.RecyclerView
+
+abstract class TreeAdapterBinder(val contentJavaClass: Class<*>) {
+    abstract val layoutId: Int
+    abstract fun provideViewHolder(itemView: View): RecyclerView.ViewHolder
+    abstract fun bindView(holder: RecyclerView.ViewHolder, node: TreeNode<*>)
+}
