+++ app/src/psi/java/org/dhis2/usescases/teiDashboard/dashboardsfragments/feedback/FeedbackContentFragment.kt
@@ -9,6 +9,8 @@ import androidx.databinding.DataBindingUtil
 import androidx.recyclerview.widget.DividerItemDecoration
 import org.dhis2.App
 import org.dhis2.R
+import org.dhis2.core.ui.tree.TreeAdapter
+import org.dhis2.core.ui.tree.TreeNode
 import org.dhis2.databinding.FragmentFeedbackContentBinding
 import org.dhis2.usescases.general.FragmentGlobalAbstract
 import org.dhis2.usescases.teiDashboard.TeiDashboardMobileActivity
@@ -89,7 +91,7 @@ class FeedbackContentFragment : FragmentGlobalAbstract(),
     }
 
     private fun setFeedbackAdapter(nodes: List<TreeNode<*>>) {
-        val adapter = FeedbackAdapter(nodes)
+        val adapter = TreeAdapter(nodes, listOf(FeedbackItemBinder(), FeedbackHelpItemBinder()))
         binding.feedbackRecyclerView.adapter = adapter
     }
